\name{cal-package}
\alias{cal-package}
\alias{cal}
\docType{package}
\title{
\packageTitle{cal}
}
\description{
\packageDescription{cal}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cal}
\packageIndices{cal}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cal}

Maintainer: \packageMaintainer{cal}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
