% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\title{Command Line Calendar}
\usage{
cal(cmd, ...)
}
\arguments{
\item{cmd}{A string in the form of Month (int/char) AND/OR Year (4 digits).
Below are the forms that are accepted:
\describe{
  \item{2017}{A full calendar for the year}
  \item{03}{A single month view}
  \item{Jan}{A single month view (case insensitive)}
  \item{January}{A single month view (case insensitive)}
  \item{2:5}{A multi month view (integers must be > 0 and < 13)}
  \item{-r 4}{A multi month view relative to the current month (curent month + n more; integers must be positive and < 13)}
  \item{-r -3:4}{A multi month view relative to the current month (curent month +/- n before/after; -r tag not required; lower must be negative; must be less than or equal to 12 months all together)}
  \item{03 2015}{A single month view for a particular year}
  \item{Jan 2015}{A single month view for a particular year (case insensitive)}
  \item{January 2015}{A single month view for a particular year (case insensitive)}
}}

\item{\ldots}{ignored.}
}
\value{
Prints calendar to the command line.
}
\description{
A command line calendar in the style of the \pkg{Unix} \pkg{cal} command
line tool though with only only flag.
}
\examples{
cal()
cal("2017")
cal("03")
cal("Jan")
cal("december")
cal("2:5")
cal("-r 4")
cal("-r -3")
cal("-r -3:4")
cal("03 2015")
cal("Jan 2015")
cal("January 2015")
}
\references{
\url{https://www.tutorialspoint.com/unix_commands/cal.htm}
}
