% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{print.cal}
\alias{print.cal}
\title{Prints a cal Object}
\usage{
\method{print}{cal}(x, ncols = 4, ...)
}
\arguments{
\item{x}{A cal Object}

\item{ncols}{The number of columns to form the print grid}

\item{\ldots}{ignored}
}
\description{
Prints a cal object
}
